requirements.txt
defusedxml==0.6.0
diff-match-patch==20181111
dj-database-url==0.5.0
Django==2.2.9
django-extensions==2.2.6
django-filter==2.2.0
django-heroku==0.3.1
django-import-export==2.0.1
django-restql==0.8.1
djangorestframework==3.11.0
drf-writable-nested==0.5.4
et-xmlfile==1.0.1
gunicorn==20.0.4
jdcal==1.4.1
Markdown==3.1.1
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.3
psycopg2==2.8.4
pyPEG2==2.15.2
pytz==2019.3
PyYAML==5.3
six==1.14.0
sqlparse==0.3.0
tablib==0.14.0
whitenoise==5.0.1
xlrd==1.2.0
xlwt==1.3.0
